Migrated upload file to the storage api
