+++ src/utils/upload-image.ts
@@ -1,8 +1,5 @@
-import * as ImagePicker from 'expo-image-picker';
 import React from 'react';
-import { getDownloadURL, ref, uploadBytesResumable } from 'firebase/storage';
-import { storage, db } from 'src/firebase/config';
-import { updateDoc, doc } from 'firebase/firestore';
+import * as ImagePicker from 'expo-image-picker';
 
 export /**
  * Lauches expo image picker for camera and changes the imageState to the URI
@@ -42,41 +39,4 @@ const pickImage = async (
     if (!result.cancelled) {
         setImageState(result.uri);
     }
-};
-
-export /**
- * Uploads file to firebase storage
- *
- * @param {string} file
- * @param {string} userID
- * @return {*}  {Promise<void>}
- */
-
-const upLoadFile = async (file: string, userID: string): Promise<void> => {
-    if (!file) return;
-    const img = await fetch(file);
-    const blobFile = await img.blob();
-    const storageRef = ref(storage, `user-profile-img/${userID}`);
-    const uploadImage = uploadBytesResumable(storageRef, blobFile);
-
-    const userRef = doc(db, 'private-user-data', userID);
-
-    uploadImage.on(
-        'state_changed',
-        (snapshot) => {
-            const percent = Math.round((snapshot.bytesTransferred / snapshot.totalBytes) * 100);
-            // useful log for seeing image being uploaded
-            console.log(percent);
-        },
-        (err) => console.warn(err),
-        () => {
-            getDownloadURL(uploadImage.snapshot.ref).then(async (url) => {
-                const data = {
-                    image: url,
-                };
-
-                await updateDoc(userRef, data);
-            });
-        },
-    );
 };
